+++ foundation-models/openai/src/test/java/com/sap/ai/sdk/foundationmodels/openai/OpenAiClientTest.java
@@ -261,7 +261,7 @@ class OpenAiClientTest extends BaseOpenAiClientTest {
       try (var stream = client.streamChatCompletionDeltas(request)) {
         assertThatThrownBy(() -> stream.forEach(System.out::println))
             .isInstanceOf(OpenAiClientException.class)
-            .hasMessage("Failed to parse response and error message: 'exceeded token rate limit'");
+            .hasMessage("Failed to parse response");
       }
 
       Mockito.verify(inputStream, times(1)).close();