+++ core/src/main/java/com/sap/ai/sdk/core/common/ClientException.java
@@ -1,6 +1,10 @@
 package com.sap.ai.sdk.core.common;
 
 import com.google.common.annotations.Beta;
+import javax.annotation.Nullable;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
 import lombok.experimental.StandardException;
 
 /**
@@ -10,4 +14,14 @@ import lombok.experimental.StandardException;
  */
 @Beta
 @StandardException
-public class ClientException extends RuntimeException {}
+public class ClientException extends RuntimeException {
+
+  /**
+   * Wraps a structured error payload received from the remote service, if available. This can be
+   * used to extract more detailed error information.
+   */
+  @Nullable
+  @Getter(onMethod_ = @Beta, value = AccessLevel.PROTECTED)
+  @Setter(onMethod_ = @Beta, value = AccessLevel.PROTECTED)
+  ClientError clientError;
+}