+++ core/src/main/java/com/sap/ai/sdk/core/common/ClientResponseHandler.java
@@ -6,38 +6,40 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.annotations.Beta;
 import io.vavr.control.Try;
-import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import java.util.Objects;
-import java.util.function.BiFunction;
+import java.util.Optional;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.apache.hc.core5.http.ClassicHttpResponse;
 import org.apache.hc.core5.http.ContentType;
 import org.apache.hc.core5.http.HttpEntity;
-import org.apache.hc.core5.http.ParseException;
-import org.apache.hc.core5.http.ProtocolException;
 import org.apache.hc.core5.http.io.HttpClientResponseHandler;
 import org.apache.hc.core5.http.io.entity.EntityUtils;
 
 /**
  * Parse incoming JSON responses and handles any errors. For internal use only.
  *
- * @param <T> The type of the response.
+ * @param <T> The type of the successful response.
  * @param <E> The type of the exception to throw.
+ * @param <R> The type of the error response.
  * @since 1.1.0
  */
 @Beta
 @Slf4j
 @RequiredArgsConstructor
-public class ClientResponseHandler<T, E extends ClientException>
+public class ClientResponseHandler<T, R extends ClientError, E extends ClientException>
     implements HttpClientResponseHandler<T> {
-  public static long time = 0;
-  @Nonnull final Class<T> responseType;
-  @Nonnull private final Class<? extends ClientError> errorType;
-  @Nonnull final BiFunction<String, Throwable, E> exceptionConstructor;
+  /** The HTTP success response type */
+  @Nonnull final Class<T> successType;
+
+  /** The HTTP error response type */
+  @Nonnull final Class<R> errorType;
+
+  /** The factory to create exceptions for Http 4xx/5xx responses. */
+  @Nonnull final ClientExceptionFactory<E, R> exceptionFactory;
 
   /** The parses for JSON responses, will be private once we can remove mixins */
   @Nonnull ObjectMapper objectMapper = getDefaultObjectMapper();
@@ -50,7 +52,7 @@ public class ClientResponseHandler<T, E extends ClientException>
    */
   @Beta
   @Nonnull
-  public ClientResponseHandler<T, E> objectMapper(@Nonnull final ObjectMapper jackson) {
+  public ClientResponseHandler<T, R, E> objectMapper(@Nonnull final ObjectMapper jackson) {
     objectMapper = jackson;
     return this;
   }
@@ -66,97 +68,105 @@ public class ClientResponseHandler<T, E extends ClientException>
   @Override
   public T handleResponse(@Nonnull final ClassicHttpResponse response) throws E {
     if (response.getCode() >= 300) {
-      buildExceptionAndThrow(response);
+      buildAndThrowException(response);
     }
-    try {
-      String value = response.getHeader("x-upstream-service-time").getValue();
-      time += Long.parseLong(value);
-    } catch (ProtocolException e) {
-      throw new RuntimeException(e);
-    }
-    return parseResponse(response);
+    return parseSuccess(response);
   }
 
   // The InputStream of the HTTP entity is closed by EntityUtils.toString
   @SuppressWarnings("PMD.CloseResource")
   @Nonnull
-  private T parseResponse(@Nonnull final ClassicHttpResponse response) throws E {
+  private T parseSuccess(@Nonnull final ClassicHttpResponse response) throws E {
     final HttpEntity responseEntity = response.getEntity();
     if (responseEntity == null) {
-      throw exceptionConstructor.apply("Response was empty.", null);
+      throw exceptionFactory.build("The HTTP Response is empty", null);
     }
-    val content = getContent(responseEntity);
+
+    val content =
+        tryGetContent(responseEntity)
+            .getOrElseThrow(e -> exceptionFactory.build("Failed to parse response entity.", e));
     try {
-      return objectMapper.readValue(content, responseType);
+      return objectMapper.readValue(content, successType);
     } catch (final JsonProcessingException e) {
-      log.error("Failed to parse response to type {}", responseType);
-      throw exceptionConstructor.apply("Failed to parse response", e);
+      log.error("Failed to parse response to type {}", successType);
+      throw exceptionFactory.build("Failed to parse response", e);
     }
   }
 
   @Nonnull
-  private String getContent(@Nonnull final HttpEntity entity) {
-    try {
-      return EntityUtils.toString(entity, StandardCharsets.UTF_8);
-    } catch (IOException | ParseException e) {
-      throw exceptionConstructor.apply("Failed to read response content", e);
-    }
+  private Try<String> tryGetContent(@Nonnull final HttpEntity entity) {
+    return Try.of(() -> EntityUtils.toString(entity, StandardCharsets.UTF_8));
   }
 
   /**
-   * Parse the error response and throw an exception.
+   * Process the error response and throw an exception.
    *
-   * @param response The response to process
+   * @param httpResponse The response to process
+   * @throws ClientException if the response is an error (4xx/5xx)
    */
   @SuppressWarnings("PMD.CloseResource")
-  public void buildExceptionAndThrow(@Nonnull final ClassicHttpResponse response) throws E {
-    val exception =
-        exceptionConstructor.apply(
-            "Request failed with status %s %s"
-                .formatted(response.getCode(), response.getReasonPhrase()),
-            null);
-    val entity = response.getEntity();
+  protected void buildAndThrowException(@Nonnull final ClassicHttpResponse httpResponse) throws E {
+
+    val entity = httpResponse.getEntity();
+
     if (entity == null) {
-      throw exception;
+      val message = getErrorMessage(httpResponse, "The HTTP Response is empty");
+      throw exceptionFactory.build(message, null);
     }
-    val maybeContent = Try.of(() -> getContent(entity));
+    val maybeContent = tryGetContent(entity);
     if (maybeContent.isFailure()) {
-      exception.addSuppressed(maybeContent.getCause());
-      throw exception;
+      val message = getErrorMessage(httpResponse, "Failed to read the response content");
+      val baseException = exceptionFactory.build(message, null);
+      baseException.addSuppressed(maybeContent.getCause());
+      throw baseException;
     }
     val content = maybeContent.get();
-    if (content.isBlank()) {
-      throw exception;
+    if (content == null || content.isBlank()) {
+      val message = getErrorMessage(httpResponse, "Empty or blank response content");
+      throw exceptionFactory.build(message, null);
     }
 
     log.error(
         "The service responded with an HTTP {} ({})",
-        response.getCode(),
-        response.getReasonPhrase());
+        httpResponse.getCode(),
+        httpResponse.getReasonPhrase());
     val contentType = ContentType.parse(entity.getContentType());
     if (!ContentType.APPLICATION_JSON.isSameMimeType(contentType)) {
-      throw exception;
+      val message = getErrorMessage(httpResponse, "The response Content-Type is not JSON");
+      throw exceptionFactory.build(message, null);
     }
 
-    parseErrorAndThrow(content, exception);
+    parseErrorResponseAndThrow(content, httpResponse);
   }
 
   /**
-   * Parse the error response and throw an exception.
+   * Parses the JSON content of an error response and throws a module specific exception.
    *
-   * @param errorResponse the error response, most likely a unique JSON class.
-   * @param baseException a base exception to add the error message to.
+   * @param content The JSON content of the error response.
+   * @param httpResponse The HTTP response that contains the error.
+   * @throws ClientException if the response is an error (4xx/5xx)
    */
-  public void parseErrorAndThrow(
-      @Nonnull final String errorResponse, @Nonnull final E baseException) throws E {
-    val maybeError = Try.of(() -> objectMapper.readValue(errorResponse, errorType));
-    if (maybeError.isFailure()) {
-      baseException.addSuppressed(maybeError.getCause());
+  protected void parseErrorResponseAndThrow(
+      @Nonnull final String content, @Nonnull final ClassicHttpResponse httpResponse) throws E {
+    val maybeClientError = Try.of(() -> objectMapper.readValue(content, errorType));
+    if (maybeClientError.isFailure()) {
+      val message = getErrorMessage(httpResponse, "Failed to parse the JSON error response");
+      val baseException = exceptionFactory.build(message, null);
+      baseException.addSuppressed(maybeClientError.getCause());
       throw baseException;
     }
+    final R clientError = maybeClientError.get();
+    val message = getErrorMessage(httpResponse, clientError.getMessage());
+    throw exceptionFactory.buildFromClientError(message, clientError);
+  }
 
-    val error = Objects.requireNonNullElse(maybeError.get().getMessage(), "");
-    val message = "%s and error message: '%s'".formatted(baseException.getMessage(), error);
-    throw exceptionConstructor.apply(message, baseException);
+  private static String getErrorMessage(
+      @Nonnull final ClassicHttpResponse httpResponse, @Nullable final String additionalMessage) {
+    val baseErrorMessage =
+        "Request failed with status %d (%s)"
+            .formatted(httpResponse.getCode(), httpResponse.getReasonPhrase());
+
+    val message = Optional.ofNullable(additionalMessage).orElse("");
+    return message.isEmpty() ? baseErrorMessage : "%s: %s".formatted(baseErrorMessage, message);
   }
 }