+++ core/src/main/java/com/sap/ai/sdk/core/model/AiPreconditionFailed.java
@@ -23,28 +23,28 @@ import java.util.Set;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
-/** InlineObject4 */
+/** AiPreconditionFailed */
 // CHECKSTYLE:OFF
-public class InlineObject4
+public class AiPreconditionFailed
 // CHECKSTYLE:ON
 {
   @JsonProperty("error")
-  private DSetError error;
+  private AiApiError error;
 
   @JsonAnySetter @JsonAnyGetter
   private final Map<String, Object> cloudSdkCustomFields = new LinkedHashMap<>();
 
-  /** Default constructor for InlineObject4. */
-  protected InlineObject4() {}
+  /** Default constructor for AiPreconditionFailed. */
+  protected AiPreconditionFailed() {}
 
   /**
-   * Set the error of this {@link InlineObject4} instance and return the same instance.
+   * Set the error of this {@link AiPreconditionFailed} instance and return the same instance.
    *
-   * @param error The error of this {@link InlineObject4}
-   * @return The same instance of this {@link InlineObject4} class
+   * @param error The error of this {@link AiPreconditionFailed}
+   * @return The same instance of this {@link AiPreconditionFailed} class
    */
   @Nonnull
-  public InlineObject4 error(@Nullable final DSetError error) {
+  public AiPreconditionFailed error(@Nullable final AiApiError error) {
     this.error = error;
     return this;
   }
@@ -52,24 +52,24 @@ public class InlineObject4
   /**
    * Get error
    *
-   * @return error The error of this {@link InlineObject4} instance.
+   * @return error The error of this {@link AiPreconditionFailed} instance.
    */
   @Nonnull
-  public DSetError getError() {
+  public AiApiError getError() {
     return error;
   }
 
   /**
-   * Set the error of this {@link InlineObject4} instance.
+   * Set the error of this {@link AiPreconditionFailed} instance.
    *
-   * @param error The error of this {@link InlineObject4}
+   * @param error The error of this {@link AiPreconditionFailed}
    */
-  public void setError(@Nullable final DSetError error) {
+  public void setError(@Nullable final AiApiError error) {
     this.error = error;
   }
 
   /**
-   * Get the names of the unrecognizable properties of the {@link InlineObject4}.
+   * Get the names of the unrecognizable properties of the {@link AiPreconditionFailed}.
    *
    * @return The set of properties names
    */
@@ -80,7 +80,7 @@ public class InlineObject4
   }
 
   /**
-   * Get the value of an unrecognizable property of this {@link InlineObject4} instance.
+   * Get the value of an unrecognizable property of this {@link AiPreconditionFailed} instance.
    *
    * @deprecated Use {@link #toMap()} instead.
    * @param name The name of the property
@@ -91,14 +91,15 @@ public class InlineObject4
   @Deprecated
   public Object getCustomField(@Nonnull final String name) throws NoSuchElementException {
     if (!cloudSdkCustomFields.containsKey(name)) {
-      throw new NoSuchElementException("InlineObject4 has no field with name '" + name + "'.");
+      throw new NoSuchElementException(
+          "AiPreconditionFailed has no field with name '" + name + "'.");
     }
     return cloudSdkCustomFields.get(name);
   }
 
   /**
-   * Get the value of all properties of this {@link InlineObject4} instance including unrecognized
-   * properties.
+   * Get the value of all properties of this {@link AiPreconditionFailed} instance including
+   * unrecognized properties.
    *
    * @return The map of all properties
    */
@@ -111,8 +112,8 @@ public class InlineObject4
   }
 
   /**
-   * Set an unrecognizable property of this {@link InlineObject4} instance. If the map previously
-   * contained a mapping for the key, the old value is replaced by the specified value.
+   * Set an unrecognizable property of this {@link AiPreconditionFailed} instance. If the map
+   * previously contained a mapping for the key, the old value is replaced by the specified value.
    *
    * @param customFieldName The name of the property
    * @param customFieldValue The value of the property
@@ -130,9 +131,9 @@ public class InlineObject4
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    final InlineObject4 inlineObject4 = (InlineObject4) o;
-    return Objects.equals(this.cloudSdkCustomFields, inlineObject4.cloudSdkCustomFields)
-        && Objects.equals(this.error, inlineObject4.error);
+    final AiPreconditionFailed aiPreconditionFailed = (AiPreconditionFailed) o;
+    return Objects.equals(this.cloudSdkCustomFields, aiPreconditionFailed.cloudSdkCustomFields)
+        && Objects.equals(this.error, aiPreconditionFailed.error);
   }
 
   @Override
@@ -144,7 +145,7 @@ public class InlineObject4
   @Nonnull
   public String toString() {
     final StringBuilder sb = new StringBuilder();
-    sb.append("class InlineObject4 {\n");
+    sb.append("class AiPreconditionFailed {\n");
     sb.append("    error: ").append(toIndentedString(error)).append("\n");
     cloudSdkCustomFields.forEach(
         (k, v) ->
@@ -163,8 +164,8 @@ public class InlineObject4
     return o.toString().replace("\n", "\n    ");
   }
 
-  /** Create a new {@link InlineObject4} instance. No arguments are required. */
-  public static InlineObject4 create() {
-    return new InlineObject4();
+  /** Create a new {@link AiPreconditionFailed} instance. No arguments are required. */
+  public static AiPreconditionFailed create() {
+    return new AiPreconditionFailed();
   }
 }